requirements.txt
requests~=2.26.0
Flask~=2.0.1
flask-restful==0.3.9
beautifulsoup4~=4.10.0
pandas==1.3.3
flask-migrate==3.1.0
flask-sqlalchemy==2.5.1
d6tstack==0.2.0
psycopg2-binary==2.9.1